requirements.txt
dotenv
djangorestframework
django
gunicorn
psycopg2